requirements.txt
discord.py
python-dotenv
# Add other dependencies here
